// Home/transaction/fin-09-invoice/model/fin09-abcde.ts
export class Fin09ABCDE {
    clinicName: string;
    clinicId: number;
    clinicCode: string;
    code: string;
    clinicTypeId: number;
    clinicTypeCode: string;
    districtId: number;
    districtCode: string;
    stateId: number;
    stateName: string;
    districtName: string;
    responseTimePenalty: number;
    repairTimePenalty: number;
    uptimePenalty: number;
    scheduledMaintenancePenalty: number;
    lateDeliveryPenalty: number;
    totalPenalty: number;
    approvalDate: any;
    approvalYear: string;
    approvalQuater: string;
    status: string;
    approvalMonth: string;
    month: any;
    remarks: any;
}